requirements.txt
langchain>=0.1.0
langgraph>=0.0.20
langchain-openai>=0.0.5
pypdf>=3.17.0
python-docx>=1.0.0
python-pptx>=0.6.21
markdown>=3.5
pydantic>=2.0.0
python-dotenv>=1.0.0
